+++ testes/ft_strlcat.c
@@ -38,8 +38,8 @@ size_t	ft_strlcat(char *dest, const char *src, size_t size)
 	size_t	src_len;
 	size_t	dest_len;
 
-	src_len = (size_t)ft_strlen(src);
-	dest_len = (size_t)ft_strlen(dest);
+	src_len = ft_strlen(src);
+	dest_len = ft_strlen(dest);
 	src_count = 0;
 	dest_count = dest_len;
 	if (size > dest_len)
