+++ testes/ft_strlcpy.c
@@ -36,7 +36,7 @@ size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 	size_t	count;
 	size_t	src_len;
 
-	src_len = (size_t)ft_strlen(src);
+	src_len = ft_strlen(src);
 	count = 0;
 	if (size > 0)
 	{
